requirements.txt
et_xmlfile==2.0.0
greenlet==3.1.1
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
py4j==0.10.9.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
python-dotenv==1.0.1
oauth2client
pyopenssl
pandas-gbq